requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-redis==5.2.0
djangorestframework==3.13.1
Faker==11.3.0
gunicorn==20.1.0
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
wrapt==1.13.3
